Refactoring controllers
